Setup coupling for esm1.6 (#33)
This change couples the fields expect by UM7.3 with CICE5, and implements spreading of meltwater from icesheets.

This coupling is limited to 0-layer thermodynamics in CICE5, due to how the fluxes from the UM are calculated and the assumptions made to calculate the fluxes.

Test results are shown in https://github.com/ACCESS-NRI/access-esm1.6-configs/issues/80

---------

Co-authored-by: Dave Bi <dhb599@gadi-login-07.gadi.nci.org.au>
Co-authored-by: Spencer Wong <88933912+blimlim@users.noreply.github.com>